Move wrapX to ol.tilegrid to break circular dependency
